% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_publication_table_utilities.R
\name{add_dotted_line}
\alias{add_dotted_line}
\title{Adds a dotted line to the publication table in specified position}
\usage{
add_dotted_line(minimal_tidy_df, all_title_text, LRT_tab, table_key)
}
\arguments{
\item{minimal_tidy_df}{is a dataframe which matches the format of the tibbles contained in the
minimal tidy dataset list.}

\item{all_title_text}{is the vector of title text in the table.}

\item{LRT_tab}{is the openxlsx object containing the data for the table.}

\item{table_key}{is the reference for the table in the table_content list.}
}
\description{
This function adds dotted lines to the publication excel table. The
positions of the dotted lines are specified by the dotted_line entry in
the table_content list.
}
