% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_QA_note_utilities.R
\name{calculate_headline_figures}
\alias{calculate_headline_figures}
\title{Calculates headline figures for UK, England, London and England outside London}
\usage{
calculate_headline_figures(lrt_tbl)
}
\arguments{
\item{lrt_tbl}{is a tibble containing the \code{last_year}, \code{this_year}, \code{diff}
and \code{percent_change} figures for a given feature (e.g. route_km_tbl) by tram system.
The tibble structure should match the tibbles stored in the list output of the
\code{\link{lrt_make_all_tables}} function.}
}
\value{
4x5 tibble containing data for UK, England, London and England outside London
}
\description{
This function takes a light rail and tram feature table with details by tram operator and
calculates the figures for UK, England, London and England outside London.
}
\examples{
calculate_headline_figures(lrt_QA_params$no_of_vehicles_tbl)

}
