% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_error_checking.R
\name{check_financial_year}
\alias{check_financial_year}
\title{Checks financial year input}
\usage{
check_financial_year(financial_year)
}
\arguments{
\item{financial_year}{is the input to be tested}
}
\description{
This function checks that the input is a string of the structure matching the financial
year format "XXXX/XX".
}
\examples{
check_financial_year("2019/20")

}
