% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_error_checking.R
\name{check_path}
\alias{check_path}
\title{Checks path input}
\usage{
check_path(path)
}
\arguments{
\item{path}{is the path inputed in to the function}
}
\description{
This function checks that the input path is a string, and that it is referencing
a directory that exists. If this is not the case, it throws an error and lets
the user know what is wrong.
}
\examples{
check_path("G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/5. RAP Data/Received Survey forms/2019-20 DLR.xlsx")

}
