% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_error_checking.R
\name{check_string}
\alias{check_string}
\title{Checks that an input is a string}
\usage{
check_string(string, string_type = "variable")
}
\arguments{
\item{string}{is the path inputed in to the function}

\item{string_type}{string to be inserted in the error message: "The ----- supplied is not a string. Make sure
it is surrounded by quotes, and that it is of length 1 (i.e. not a vector).". The default value is "variable".}
}
\description{
This function checks that the input is a string, and if not, throws an error message
}
\examples{
check_string("Hello world")

}
