% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_error_checking.R
\name{check_year_gap}
\alias{check_year_gap}
\title{Checks year gaps input}
\usage{
check_year_gap(year_gap)
}
\arguments{
\item{year_gap}{is the input to be tested}
}
\description{
This function checks that the input is an integer between 1 and 12.
}
\examples{
check_year_gap(10)

}
