% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_QA_note_generate.R
\name{generate_lrt_QA_note}
\alias{generate_lrt_QA_note}
\title{Generates QA note for the light rail and tram publication}
\usage{
generate_lrt_QA_note(survey_sent_path, survey_response_path,
  save_QA_note_path, publication_fin_year, email_response_path = NULL,
  QA_author = "Rose McNally", QA_date = NULL)
}
\arguments{
\item{survey_sent_path}{is a string containing the path of the survey sent folder. This folder
must contain the sent survey forms and nothing else. The path should not end with slashes. For Windows
paths, each backslash must be changed to either a forward slash "/" or two backslashes "\\\\".}

\item{survey_response_path}{is a string containing the path of the survey responses folder. This folder
must contain the survey responses and nothing else. The path should not end with slashes. For Windows
paths, each backslash must be changed to either a forward slash "/" or two backslashes "\\\\".}

\item{save_QA_note_path}{is a string containing the path of the folder in which you would
like to save the QA note file. The path should not end with slashes. For Windows paths, each
backslash must be changed to either a forward slash "/" or two backslashes "\\\\".}

\item{publication_fin_year}{is a string containing the financial year to which the survey data
refers. For example: "2019/20".}

\item{email_response_path}{is a string containing the path of an excel spreadsheet. The
spreadsheet should consist of one "name" column which contains the names of the light rail or
tram system (must match the names in the survey forms) and one "comments" column containing the
relevant information from the emails. For Windows paths, each backslash must be changed to
either a forward slash "/" or two backslashes "\\\\". If there are no email responses to include
in the QA note then leave the field blank, the default value is null.}

\item{QA_author}{is a string containg the name of the author, which will appear on the QA note. If
this is not set, the default is "Rose McNally".}

\item{QA_date}{is a string containing the date which shall appear at the top of the QA note. It should
be in the form "15 May 2020". If this is not set, the default is the date on which the QA note is
generated.}
}
\description{
This function reads the excel files containing all of the relevant survey data from
operators. It then generates a QA note and saves it as an html file.
}
\examples{
generate_lrt_QA_note("G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/2. Survey set up/Survey forms issued",
                     "G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/3. Survey/Received Survey forms",
                     "G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/6. Validation",
                     "2019/20",
                     "G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/3. Survey/Survey response emails.xlsx")

}
