% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_generate_publication_charts.R
\name{generate_publication_charts}
\alias{generate_publication_charts}
\title{Produces Charts 1-4 for the light rail and tram publication}
\usage{
generate_publication_charts(minimal_tidy_dataset_path,
  save_publication_charts_path, publication_fin_year,
  year_gaps_chart_1 = 6)
}
\arguments{
\item{minimal_tidy_dataset_path}{is a string containing the file path of the Minimal Tidy Dataset
excel file. For Windows paths, each backslash must be changed to either a forward slash "/"
or two backslashes "\\\\".}

\item{save_publication_charts_path}{is a string containing the file path of the folder in which
you would like to save the folder of charts. For Windows paths, each backslash
must be changed to either a forward slash "/" or two backslashes "\\\\".}

\item{publication_fin_year}{is a string containing the financial year to which the survey data
refers. For example: "2019/20".}

\item{year_gaps_chart_1}{is an integer specifying the difference in years between each x-axis label.
For example: the number of years between 1983/84 and 1989/90 is 6. The default is set to 6, the maximum is
12. If the number of years is not easily divisible along the axis, all but the final year will be spaced
evenly along the axis.}
}
\description{
This function reads the data in the minimal tidy dataset, produces the four charts needed for
the publication and then saves them as jpegs in the file location given in the arguments.
}
\examples{
generate_publication_charts("G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/7. Minimal Tidy Dataset/2020_minimal_tidy_dataset.xlsx",
                            "G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0002 Publication/20-05 Light Rail Statistics",
                            "2019/20")
}
