% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_generate_publication_figures.R
\name{generate_publication_figures}
\alias{generate_publication_figures}
\title{Produces an HTML document with all the figures needed for the publication}
\usage{
generate_publication_figures(minimal_tidy_dataset_path,
  save_publication_figures_path, publication_fin_year,
  figures_author = NULL)
}
\arguments{
\item{minimal_tidy_dataset_path}{is a string containing the file path of the Minimal Tidy Dataset
excel file. For Windows paths, each backslash must be changed to either a forward slash "/"
or two backslashes "\\\\".}

\item{save_publication_figures_path}{is a string containing the file path of the folder in which
you would like to save the HTML file. For Windows paths, each backslash
must be changed to either a forward slash "/" or two backslashes "\\\\".}

\item{publication_fin_year}{is a string containing the financial year to which the survey data
refers. For example: "2019/20".}
}
\description{
This function reads the data in the minimal tidy dataset and produces an HTML document with
RMarkdown which contains all the figures needed for the light rail and tram publication.
}
\examples{
generate_publication_figures("G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/7. Minimal Tidy Dataset/2020_minimal_tidy_dataset.xlsx",
                             "G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0002 Publication/20-05 Light Rail Statistics",
                             "2019/20")
}
