% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_generate_publication_tables.R
\name{generate_publication_tables}
\alias{generate_publication_tables}
\title{Converts the minimal tidy dataset into publication tables in excel}
\usage{
generate_publication_tables(minimal_tidy_dataset_path,
  save_publication_tables_path, publication_fin_year, publication_date)
}
\arguments{
\item{minimal_tidy_dataset_path}{is a string containing the file path of the Minimal Tidy Dataset
excel file. For Windows paths, each backslash must be changed to either a forward slash "/"
or two backslashes "\\\\".}

\item{save_publication_tables_path}{is a string containing the file path of the folder in which
you would like to save the folder of excel files. For Windows paths, each backslash
must be changed to either a forward slash "/" or two backslashes "\\\\".}

\item{publication_fin_year}{is a string containing the financial year to which the survey data
refers. For example: "2019/20".}

\item{publication_date}{is a string containing the date of the publication. For example:
"25th June 2020".}
}
\description{
This function reads the data in the minimal tidy dataset and converts each table into
a fully formatted, publication-ready excel file. The excel files are saved in a new folder
in the directory specified in the arguments.
}
\examples{
generate_publication_tables("G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/7. Minimal Tidy Dataset/2020_minimal_tidy_dataset.xlsx",
                            "G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0002 Publication/20-05 Light Rail Statistics/Tables",
                            "2019/20",
                            "25th June 2020")
}
