% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_generate_survey_forms.R
\name{generate_survey_forms}
\alias{generate_survey_forms}
\title{Generates new survey forms}
\usage{
generate_survey_forms(financial_year, survey_deadline,
  last_surveys_folder_path, save_surveys_path)
}
\arguments{
\item{financial_year}{is a string containing the financial year to which the survey
refers. For example "2020/21".}

\item{survey_deadline}{is a string containing the date by which the surveys must be
returned. It should be in the form "Day Month Year". For example: "5th May 2020".}

\item{last_surveys_folder_path}{is a string containing the path of the surveys folder for last year.
This folder must contain the validated survey forms and nothing else. The path should not end with slashes.
For Windows paths, each backslash must be changed to either a forward slash "/" or two backslashes "\\\\".}

\item{save_surveys_path}{is a string containing the path to the directory in which you
would like the survey forms to be saved. A folder with the forms will be created in the
directory given. For Windows paths, each backslash must be changed to either a forward slash "/"
or two backslashes "\\\\".}
}
\description{
This function generates the surveys forms for the next light rail and tram survey
and publication.
}
\details{
This function takes the blank survey form saved in inst/extdata and substitutes in
the correct dates from the arguments. For each operator it fills in the last year
figures from last year's validated survey forms (which have the finalised figures in)
and saves the excel form in the specified location.
}
\examples{
generate_survey_forms("2020/21",
                      "Example date 2021",
                      "G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/3. Survey/Validated survey forms",
                      "G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2021/2. Survey set up")

}
