% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_QA_note_utilities.R
\name{get_previous_fin_year}
\alias{get_previous_fin_year}
\title{Gets previous financial year from given financial year}
\usage{
get_previous_fin_year(fin_year)
}
\arguments{
\item{fin_year}{is a string of a financial year, in the form "XXXX/XX".}
}
\value{
Output is a string of the previous financial year in the form "XXXX/XX".
}
\description{
This function takes a string of a financial year, e.g. "2019/20" and returns the previous
financial year, e.g. "2018/19".
}
\examples{
get_previous_fin_year("2019/20")
get_previous_fin_year("2013/14")

}
