% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_general_utilities.R
\name{get_separate_years}
\alias{get_separate_years}
\title{Gets both full years from financial year}
\usage{
get_separate_years(financial_year)
}
\arguments{
\item{financial_year}{is a string of the financial year, in
the form "XXXX/XX".}
}
\value{
a vector containing the two years. The structure is
c(year_1, year_2).
}
\description{
This function reads the financial year string and returns
the two separate years in full, as elements of a vector.
}
\examples{
get_separate_years("2019/20")

}
