% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_QA_note_utilities.R
\name{inc_or_dec}
\alias{inc_or_dec}
\title{Checks whether a percentage change is an increase or decrease}
\usage{
inc_or_dec(lrt_tbl_row, previous_fin_year, currency = FALSE,
  commas = FALSE, km = FALSE)
}
\arguments{
\item{lrt_tbl_row}{is a row from a light rail and tram feature table.}

\item{previous_fin_year}{is a string containing the previous financial year, e.g. "2018/19".}

\item{currency}{boolean, if set to TRUE, the figures will be printed with a pound sign. Default
is FALSE.}

\item{commas}{boolean, if set to TRUE, the figures will be printed with commas. Default
is FALSE.}

\item{km}{boolean, if set to TRUE, the figures will be printed with "km" at the end. Default
is FALSE.}
}
\value{
Output is a string.
}
\description{
This function takes a row from an light rail and tram feature table, checks whether
the percent_change is positive, negative, or equal to zero, and outputs a string
describing the change. To be used in rmarkdown to fill in text about change since last year.
}
\examples{
inc_or_dec(lrt_QA_params$no_of_vehicles_tbl[1,])
inc_or_dec(lrt_QA_params$cons_eld_dis[6,])


}
