% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_QA_check_last_year_fields.R
\name{last_year_check}
\alias{last_year_check}
\title{Check whether "last year" data has been changed by operators}
\usage{
last_year_check(survey_sent, survey_response)
}
\arguments{
\item{survey_sent}{is a tibble containing the data from the survey forms sent out to operators.
The tibble structure should match the output of the \code{\link{read_lrt_folder}} function.}

\item{survey_response}{is a tibble containing the data from the survey forms returned by operators.
The tibble structure should match the output of the \code{\link{read_lrt_folder}} function.}
}
\value{
A tibble with the details of any "last year" values which were changed. If none were changed, returns
NULL.
}
\description{
This function compares the data from the surveys sent and the survey responses to check
whether the "last year" data that was sent in the forms has been changed by the operators.
}
\examples{
last_year_check(read_lrt_folder("G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/5. RAP Data/Survey forms issued"),
                read_lrt_folder("G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/5. RAP Data/Received Survey forms"))

}
