% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_QA_check_last_year_fields.R
\name{last_year_only}
\alias{last_year_only}
\title{Filters survey tibble to give last year entries only}
\usage{
last_year_only(survey_tibble)
}
\arguments{
\item{survey_tibble}{is a tibble containing the data from the survey forms.
The tibble structure should match the output of the \code{\link{read_lrt_folder}} function.}
}
\value{
Filtered tibble with "last year" entries only.
}
\description{
This function reduces the survey form tibble down to only the rows where year = "last_year",
and drops the columns which don't have a "last_year" field in the form (changes_to_fleet,
basis, source_details, description, comments).
}
\examples{
last_year_only(read_lrt_folder("G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/5. RAP Data/Received Survey forms"))

}
