% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_QA_note_plots.R
\name{lrt_bar_chart}
\alias{lrt_bar_chart}
\title{Plots bar chart}
\usage{
lrt_bar_chart(lrt_tbl, previous_fin_year, publication_fin_year,
  lrt_colour = "#d25f15")
}
\arguments{
\item{lrt_tbl}{is a tibble containing the \code{last_year}, \code{this_year}, \code{diff}
and \code{percent_change} figures for a given feature (e.g. route_km_tbl) by tram system.
The tibble structure should match the tibbles stored in the list output of the
\code{\link{lrt_make_all_tables}} function.}

\item{previous_fin_year}{is a string containing the financial year before the one to which the
survey data refers. For example: "2018/19".}

\item{publication_fin_year}{is a string containing the financial year to which the survey data
refers. For example: "2019/20".}

\item{plot_colour}{is the HEX colour code for the plot, default is orange.}
}
\value{
Dynamic bar chart of percentage change by network.
}
\description{
This function takes a light rail and tram feature table, a HEX colour code, and two
financial years and plots a bar chart of percentage change for each network.
}
\examples{
lrt_bar_chart(params$no_of_stops_tbl, previous_fin_year, publication_fin_year)

}
