% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_QA_note_plots.R
\name{lrt_rmd_table}
\alias{lrt_rmd_table}
\title{Displays table in rmarkdown}
\usage{
lrt_rmd_table(lrt_tbl, previous_fin_year, publication_fin_year,
  lrt_caption = "")
}
\arguments{
\item{lrt_tbl}{is a tibble containing the \code{last_year}, \code{this_year}, \code{diff}
and \code{percent_change} figures for a given feature (e.g. route_km_tbl) by tram system.
The tibble structure should match the tibbles stored in the list output of the
\code{\link{lrt_make_all_tables}} function.}

\item{previous_fin_year}{is a string containing the financial year before the one to which the
survey data refers. For example: "2018/19".}

\item{publication_fin_year}{is a string containing the financial year to which the survey data
refers. For example: "2019/20".}

\item{lrt_caption}{is a string that appears above the table. The default is an empty string.}
}
\value{
A table in rmarkdown.
}
\description{
This function takes a light rail and tram feature table, a caption, and two financial years
and displays a table in rmarkdown.
}
\examples{
lrt_rmd_table(params$no_of_stops_tbl, previous_fin_year, publication_fin_year, lrt_caption = "Number of tram stops")

}
