% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_make_publication_table.R
\name{make_publication_table}
\alias{make_publication_table}
\title{Converts one minimal tidy dataset table into a publication table in excel}
\usage{
make_publication_table(minimal_tidy_table, minimal_tidy_table_name,
  publication_tables_folder_path, publication_date)
}
\arguments{
\item{minimal_tidy_table}{is a tibble which matches the format of those contained in the
minimal tidy dataset list.}

\item{minimal_tidy_table_name}{is a string containing the name of the minimal tidy table,
for example "LRT0104_passenger_miles".}

\item{publication_date}{is a string containing the date of the publication, for example
"25th June 2020". The date must be given in this format.}

\item{save_publication_tables_path}{is a string containing the file path of the folder in which
you would like to save the excel file. For Windows paths, each backslash
must be changed to either a forward slash "/" or two backslashes "\\\\".}
}
\description{
This function reads the data in one minimal tidy dataset table and converts it into
a fully formatted, publication-ready excel file. The excel file is saved in the folder
specified in the arguments. This function is called in the \code{\link{generate_publication_tables}}
function. The function uses constants from the \code{publication_table_constants.R} file
to determine the content of the table.
}
