% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_make_survey_form.R
\name{make_survey_form}
\alias{make_survey_form}
\title{Makes a new survey form}
\usage{
make_survey_form(wb, operator_name, financial_year, last_survey_forms,
  surveys_folder_path)
}
\arguments{
\item{wb}{is an openxlsx workbook object, containing the data for the blank excel survey
form with the correct dates substituted in.}

\item{operator_name}{is a string containing the name of the operator to which the survey form
ia addressed. It must be the same as that which appears in the minimal tidy dataset headers.}

\item{financial_year}{is a string containing the financial year to which the survey
refers. For example "2020/21".}

\item{last_survey_forms}{is a tibble containing the data entered into last year's validated
survey forms. It must match the output of \code{read_lrt_folder}.}

\item{surveys_folder_path}{is a string containing the path to the directory in which you
would like the survey form to be saved. For Windows paths, each backslash must be changed
to either a forward slash "/" or two backslashes "\\\\".}
}
\description{
This function makes a new survey form for next year's survey and publication.
}
\examples{
make_survey_form(wb,
                 "Blackpool Tramway",
                 financial_year,
                 last_survey_forms,
                 surveys_folder_path)

}
