% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_QA_note_utilities.R
\name{na_message}
\alias{na_message}
\title{Checks for NAs and returns message if they are present}
\usage{
na_message(lrt_tbl)
}
\arguments{
\item{lrt_tbl}{is a tibble containing the \code{last_year}, \code{this_year}, \code{diff}
and \code{percent_change} figures for a given feature (e.g. route_km_tbl) by tram system.
The tibble structure should match the tibbles stored in the list output of the
\code{\link{lrt_make_all_tables}} function.}
}
\value{
Output is a list containing the NA message strings. If there are none, the ouput
is an empty list.
}
\description{
This function takes a light rail and tram feature table and returns a message if any of
the entries are NA.
}
\examples{
na_message(passenger_receipts_tbl)

}
