% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_read_lrt_files.R
\name{read_email_response}
\alias{read_email_response}
\title{Reads email comments file}
\usage{
read_email_response(email_response_path)
}
\arguments{
\item{email_response_path}{is a string containing the path of an excel spreadsheet. The
spreadsheet should consist of one "name" column which contains the names of the light rail or
tram system (must match the names in the survey forms) and one "comments" column containing the
relevant information from the emails. For Windows paths, each backslash must be changed to
either a forward slash "/" or two backslashes "\\\\".}
}
\value{
A tibble containing the email responses.
}
\description{
If you wish to include any details given in the email response which were not included
in the survey form, this funciton will read an excel file with the details in.
}
\examples{
read_email_response("G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/5. RAP Data/Survey response emails.xlsx")

}
