% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_read_lrt_files.R
\name{read_gdp_deflator}
\alias{read_gdp_deflator}
\title{Reads and tidys the GDP Deflator excel file}
\usage{
read_gdp_deflator(gdp_deflator_path, publication_fin_year)
}
\arguments{
\item{gdp_deflator_path}{is a string containing the file path of the GDP Deflator
excel file. For Windows paths, each backslash must be changed to either a forward slash "/"
or two backslashes "\\\\".}

\item{publication_fin_year}{is a string containing the financial year to which the survey data
refers. For example: "2019/20".}
}
\value{
A tibble containing the GDP Deflator values from 1983/84 to the publication financial year.
}
\description{
This function reads the GDP Deflator excel file and returns
a tidied tibble containing the useful data.
}
\examples{
read_gdp_deflator("G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/4. Data Entry and Table Compilation/ONS/GDP_Deflators_Qtrly_National_Acdplyr::counts_March_2020.xlsx",
                  "2019/20")

}
