% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_read_lrt_files.R
\name{read_lrt_file}
\alias{read_lrt_file}
\title{Reads a Light Rail and Tram response form}
\usage{
read_lrt_file(survey_path)
}
\arguments{
\item{survey_path}{is a string containing the file path of the survey form.
For Windows paths, each backslash must be changed to either a forward slash "/"
or two backslashes "\\\\".}
}
\value{
If the file location and format of the survey form are both
correct, the output will be a 2x17 tibble containing all the data entries
in the survey form. This does not include the contact details.
}
\description{
This function reads the Light Rail and Tram survey form and returns a
tibble containing the data from the form.
}
\examples{
read_lrt_file("G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/5. RAP Data/Received Survey forms/2019-20 DLR.xlsx")

}
