% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_read_lrt_files.R
\name{read_lrt_folder}
\alias{read_lrt_folder}
\title{Loops through Light Rail and Tram survey folder and return full dataset}
\usage{
read_lrt_folder(survey_folder_path)
}
\arguments{
\item{survey_folder_path}{is a string containing the path of the surveys folder. This folder
must contain the survey forms and nothing else. The path should not end with slashes. For Windows
paths, each backslash must be changed to either a forward slash "/" or two backslashes "\\\\".}
}
\value{
If the folder location and format are both correct, the output will be a 2dx17
tibble (where d is the number of survey responses in the folder) containing all the data entries
in each survey form. This does not include the contact details.
}
\description{
This function loops through a folder with the light rail and tram survey forms in
and returns a tibble containing the data from each survey form.
}
\examples{
read_lrt_folder("G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/5. RAP Data/Received Survey forms")

}
