% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_read_lrt_files.R
\name{read_min_tidy_dataset}
\alias{read_min_tidy_dataset}
\title{Reads the Minimal Tidy Dataset excel file}
\usage{
read_min_tidy_dataset(min_tidy_dataset_path)
}
\arguments{
\item{min_tidy_dataset_path}{is a string containing the file path of the Minimal Tidy Dataset
excel file. For Windows paths, each backslash must be changed to either a forward slash "/"
or two backslashes "\\\\".}
}
\value{
A list of tibbles containing the light rail and tram time series data
}
\description{
This function reads the light rail and tram Minimal Tidy Dataset excel file and returns
a list of tibbles containing the timeseries data.
}
\examples{
read_min_tidy_dataset("G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/7. Minimal Tidy Dataset/2019_minimal_tidy_dataset.xlsx")

}
