% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_read_lrt_files.R
\name{read_population_mye}
\alias{read_population_mye}
\title{Reads and tidies the population mid-year estimates excel file}
\usage{
read_population_mye(population_mye_path, publication_fin_year)
}
\arguments{
\item{population_mye_path}{is a string containing the file path of the population mid-year estimates
excel file. For Windows paths, each backslash must be changed to either a forward slash "/"
or two backslashes "\\\\".}

\item{publication_fin_year}{is a string containing the financial year to which the survey data
refers. For example: "2019/20".}
}
\value{
A tibble containing the population mid-year estimates value for each area for the most recent
year in the excel file.
}
\description{
This function reads the population mid-year estimates excel file and returns
a tidied tibble containing the useful data.
}
\examples{
read_population_mye("G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/4. Data Entry and Table Compilation/ONS/ukmidyearestimates20182019ladcodes.xls",
                    "2019/20")

}
