% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_general_utilities.R
\name{round_up}
\alias{round_up}
\title{Rounding function which rounds .5 up}
\usage{
round_up(x, digits = 0)
}
\arguments{
\item{x}{is the number to be rounded.}

\item{digits}{is the number of digits to round to. If digits = -1 then it will round to the
nearest 10.}
}
\description{
This function rounds numbers as \code{\link{round}} does, except it does not use the
round-to-even rule, so 0.5 is rounded to 1, not 0.
}
\examples{
round_up(2.5, 0)

}
