% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_publication_table_utilities.R
\name{set_footer_row_heights}
\alias{set_footer_row_heights}
\title{Sets the heights of cells with long footers}
\usage{
set_footer_row_heights(table_key, all_title_text, headers_all,
  minimal_tidy_df, footer_text, LRT_tab)
}
\arguments{
\item{table_key}{is the reference for the table in the table_content list.}

\item{all_title_text}{is the vector of title text in the table.}

\item{headers_all}{is the vecotr containing the table headers.}

\item{minimal_tidy_df}{is a dataframe which matches the format of the tibbles contained in the
minimal tidy dataset list.}

\item{footer_text}{is the vector containing all of the table footers.}

\item{LRT_tab}{is the openxlsx object containing the data for the table.}
}
\description{
This function sets the heights of the cells containing the footers which are longer than the
width of the table.
}
