% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_update_minimal_tidy_dataset.R
\name{update_minimal_tidy_dataset}
\alias{update_minimal_tidy_dataset}
\title{Updates the minimal tidy dataset and saves it as a new excel file}
\usage{
update_minimal_tidy_dataset(min_tidy_dataset_path, survey_folder_path,
  gdp_deflator_path, population_mye_path, publication_fin_year,
  save_min_tidy_dataset_path)
}
\arguments{
\item{min_tidy_dataset_path}{is a string containing the file path of the Minimal Tidy Dataset
excel file from last year. For Windows paths, each backslash must be changed to either a forward slash "/"
or two backslashes "\\\\".}

\item{survey_folder_path}{is a string containing the path of the surveys folder. This folder
must contain the survey forms and nothing else. The path should not end with slashes. For Windows
paths, each backslash must be changed to either a forward slash "/" or two backslashes "\\\\".}

\item{gdp_deflator_path}{is a string containing the file path of the GDP Deflator
excel file. For Windows paths, each backslash must be changed to either a forward slash "/"
or two backslashes "\\\\".}

\item{population_mye_path}{is a string containing the file path of the population mid-year estimates
excel file. For Windows paths, each backslash must be changed to either a forward slash "/"
or two backslashes "\\\\".}

\item{publication_fin_year}{is a string containing the financial year to which the survey data
refers. For example: "2019/20".}

\item{save_min_tidy_dataset_path}{is a string containing the file path of the folder in which
you would like to save the updated minimal tidy dataset file. For Windows paths, each backslash
must be changed to either a forward slash "/" or two backslashes "\\\\".}
}
\description{
This function updates the minimal tidy dataset with the most recent data from the survey, and then
saves the updated data as a new excel file.
}
\examples{
update_minimal_tidy_dataset("G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/7. Minimal Tidy Dataset/2019_minimal_tidy_dataset.xlsx",
                            "G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/3. Survey/Validated survey forms",
                            "G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/4. Data Entry and Table Compilation/ONS/GDP_Deflators_Qtrly_National_Accounts_March_2020.xlsx",
                            "G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/4. Data Entry and Table Compilation/ONS/ukmidyearestimates20182019ladcodes.xls",
                            "2019/20",
                            "G:/AFP/RLTDAll/STS/003 BLT/003 LIGHT RAIL/0001 Data/2020/7. Minimal Tidy Dataset")

}
